Add files via upload
This add-on adds ring loops with a gap between them for easier decimation.
The function is listed under "select" in edit mode. But i recommend assigning it the shortcut alt+R. As it isn't used in edit mode.
How it works:
1. Select a single edge.
2. Click the button.
3. You can press control+X to dissolve your selection.

Alternatively the problem can also be solved without this add-on.
1. select an edge
2. click select>loops>ring loop.
3. click select>checker deselect.
4. click select>loops>edge loop.
5. You can press control+X to dissolve your selection.
and you got the same result.